New features and Compatability with Game version 1.3.6f1

+++ InfoLoom/Systems/TradeCostData/TradeCostDomain/ResourceTradeCost.cs
@@ -7,7 +7,7 @@ namespace InfoLoomTwo.Systems.TradeCostData.TradeCostDomain
         public int Count;
         public float BuyCost;
         public float SellCost;
-        public int ImportAmount;
-        public int ExportAmount;
+        public float ImportAmount;
+        public float ExportAmount;
     }
 }
